[FIX] website: resolve page getting crash issue
Steps to Reproduce :
1. Activate the developer mode
2. Create a menu from the Site ---> Menu Editor option
3. Go to Configurations ---> Redirects
    ​3.1. Select Action as : '301 Moved Permanently'
    ​3.2. URL From : '/<url_added_at_time_Menu_creation'
    ​3.3. URL To : '#'
4. Go to the Home page
5. Click on the Menu created twice
6. Traceback Occurs

Reason:
When the path of 'url_to' and 'url_from' are same, the redirection loop
is created  and page is getting crashed.

This commit is resolving page crash issue by checking the path of
'url_to' and 'url_from' in '_serve_fallback' method, where redirection
is happening. If path is same, then we will not redirect; instead, we
will return from this method. If 'url_to' starts with '#' or '?',then in
this case we are showing validation error to the user.

task-3984211

closes odoo/odoo#213559

X-original-commit: 64925e510afe2fe834fc800ba77df34b518205e6
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Co-authored-by: Divyesh Vyas <divy@odoo.com>