[FIX] hr_holidays: Add tests to use correct unit for time off types
steps to reproduce:
1. create an accrual plan that is accrued in hours
(`added_value_type = 'hours'`)
2. create a time off type that can only be taken in days
3. create an accrual allocation with the plan and that time off
type

The calculation in the dashboard shows the amount of hours as
days.

The issue is that the amount of time off allocated is shown
for the time off type request which is not true in this case.

Note: Same fix was done a while ago: https://github.com/odoo/odoo/commit/f9c5502cb7beb81149683f68666459b09c7df891
But the tests remain valid.

closes odoo/odoo#206532

Task: 4076800
X-original-commit: 81817d0fa667ce963782b0499aa9ca7363532db5
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>