[PERF] mail: Disable prefetching of unneeded fields
Modifications are made around read, filtered, list comprehension.

The read function usually does not prefetech fields other than
stated. But in this case, since the fields are related fields
from another model, The fields of the other model are prefeteched.
This includes the body which can be very big in size and cause an
out of memory error.

Filtered does not need other fields and keeps them unnecessarily
in cache.

List comprehension where it has a reference to specific field inside
the model also triggers the prefetecher.

Benchmark:
|        |Number of queries| SQL time| Python time||
|-------|----------------------------|----------------|---------------|-|
|with prefetch| 82| 1.887 | 7.898| Out of memory|
|Without prefetch| 120| 0.385 | 5.973| |

The benchmark with the prefetch was done locally with
increasing the memory limit to get the data.

closes odoo/odoo#211075

X-original-commit: a58b6ce14124f64af2ac82da01f7cd17b5bdb8b5
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>