[FIX] l10n_tr_nilvera_einvoice: amount in text
This commit will add the amount residual in text in the note of the xml we sent
to nilvera. The text need to be in uppercase.
If the invoice is in another currency than Turkish Lira, we have to add two
notes one for the amount in turkish lira and one in the other currency
Also when the integer part or fractional part is zero it should be zero in
turkish.

task-4518269

closes odoo/odoo#217910

X-original-commit: a521586ee57ee0e9b3c041d0ea5935ec9892eddd
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>